trojkat
-stworzyc nowy model
-klasa zwracana na widok

+++ Test/Test/Models/Trojkat.cs
@@ -2,20 +2,21 @@
 {
     public class Trojkat
     {
-        public int a {  get; set; }
-        public int b { get; set; }
-        public int c { get; set; }
+        public double a {  get; set; }
+        public double b { get; set; }
+        public double c { get; set; }        
+
         /*public Trojkat(int _a, int _b, int _c)
         {
             a = _a; 
             b = _b; 
             c = _c;
         }*/
-        public int PP()
+        public double PP()
         {
             return (a * b) / 2;
         }
-        public int Obw()
+        public double Obw()
         {
             return a + b + c;
         }
